+++ CHANGES.md
@@ -3,6 +3,7 @@
 ## Unreleased - 2020-xx-xx
 ### Changed
 * Bumped `rand` to `0.8`
+* Update `rust-tls` to `0.19.0`
 
 ### Fixed
 * added the actual parsing error to `test::read_body_json` [#1812]