+++ CHANGES.md
@@ -3,6 +3,7 @@
 ## Unreleased - 2020-xx-xx
 ### Changed
 * Bumped `rand` to `0.8`
+* Rename `Handler` to `HandlerService` and rename `Factory` to `Handler`. [#1852]
 
 ### Fixed
 * added the actual parsing error to `test::read_body_json` [#1812]