+++ CHANGES.md
@@ -9,8 +9,7 @@
 * added the actual parsing error to `test::read_body_json` [#1812]
 
 [#1812]: https://github.com/actix/actix-web/pull/1812
-
-
+[#1852]: https://github.com/actix/actix-web/pull/1852
 
 ## 3.3.2 - 2020-12-01
 ### Fixed